feat: refactored order-view util functions to shared directory
