+++ shared/index.ts
@@ -2,3 +2,4 @@ export * from './amount-util'
 export * from './transaction-util'
 export * from './event-util'
 export * from './contract-util'
+export * from './order-view-util'
